+++ security/src/main/java/gohigher/oauth2/handler/AuthenticationSuccessHandler.java
@@ -2,14 +2,20 @@ package gohigher.oauth2.handler;
 
 import java.io.IOException;
 import java.nio.charset.StandardCharsets;
+import java.util.Date;
 
 import org.springframework.beans.factory.annotation.Value;
+import org.springframework.http.HttpHeaders;
+import org.springframework.http.ResponseCookie;
 import org.springframework.security.core.Authentication;
 import org.springframework.security.oauth2.core.user.OAuth2User;
 import org.springframework.security.web.authentication.SimpleUrlAuthenticationSuccessHandler;
 import org.springframework.stereotype.Component;
 import org.springframework.web.util.UriComponentsBuilder;
 
+import gohigher.auth.support.JwtProvider;
+import gohigher.auth.support.RefreshTokenCookieProvider;
+import gohigher.auth.support.TokenType;
 import gohigher.global.exception.GlobalErrorCode;
 import gohigher.global.exception.GoHigherException;
 import jakarta.servlet.http.HttpServletRequest;
@@ -20,11 +26,16 @@ public class AuthenticationSuccessHandler extends SimpleUrlAuthenticationSuccess
 
 	private final String redirectUri;
 	private final String tokenRequestUri;
+	private final JwtProvider jwtProvider;
+	private final RefreshTokenCookieProvider refreshTokenCookieProvider;
 
 	public AuthenticationSuccessHandler(@Value("${oauth2.success.redirect_uri}") String redirectUri,
-		@Value("${token.request.uri}") String tokenRequestUri) {
+		@Value("${token.request.uri}") String tokenRequestUri, JwtProvider jwtProvider,
+		RefreshTokenCookieProvider refreshTokenCookieProvider) {
 		this.redirectUri = redirectUri;
 		this.tokenRequestUri = tokenRequestUri;
+		this.jwtProvider = jwtProvider;
+		this.refreshTokenCookieProvider = refreshTokenCookieProvider;
 	}
 
 	@Override
@@ -35,7 +46,12 @@ public class AuthenticationSuccessHandler extends SimpleUrlAuthenticationSuccess
 		Long userId = oAuth2User.getAttribute("userId");
 		String role = getRole(oAuth2User);
 
-		getRedirectStrategy().sendRedirect(request, response, createTargetUrl(userId, role));
+		Date today = new Date();
+		String accessToken = jwtProvider.createToken(userId, today, TokenType.ACCESS);
+		String refreshToken = jwtProvider.createToken(userId, today, TokenType.REFRESH);
+
+		addRefreshTokenCookie(response, refreshToken);
+		getRedirectStrategy().sendRedirect(request, response, createTargetUrl(accessToken, role));
 	}
 
 	private String getRole(OAuth2User oAuth2User) {
@@ -45,12 +61,17 @@ public class AuthenticationSuccessHandler extends SimpleUrlAuthenticationSuccess
 			.getAuthority();
 	}
 
-	private String createTargetUrl(Long userId, String role) {
+	private String createTargetUrl(String accessToken, String role) {
 		return UriComponentsBuilder.fromUriString(redirectUri + tokenRequestUri)
-			.queryParam("userId", userId)
+			.queryParam("accessToken", accessToken)
 			.queryParam("role", role)
 			.build()
 			.encode(StandardCharsets.UTF_8)
 			.toUriString();
+	}
+
+	private void addRefreshTokenCookie(HttpServletResponse response, String refreshToken) {
+		ResponseCookie responseCookie = refreshTokenCookieProvider.create(refreshToken);
+		response.addHeader(HttpHeaders.SET_COOKIE, responseCookie.toString());
 	}
 }
